+++ desktop/src/com/angian/libgdx/pipes/desktop/DesktopLauncher.java
@@ -1,8 +1,8 @@
 package com.angian.libgdx.pipes.desktop;
 
+import com.angian.libgdx.pipes.PipesGame;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.angian.libgdx.pipes.PipesGame;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {